+++ packages/yak_error_handler/lib/src/all.dart
@@ -1,3 +1,4 @@
 export 'custom/all.dart' hide WrongErrorHandler;
 export 'error/all.dart';
 export 'exception/all.dart';
+export 'extension/all.dart';