+++ packages/yak_error_handler/test/custom/wrong_handler_test.dart
@@ -1,5 +1,3 @@
-import 'dart:math';
-
 import 'package:test/test.dart';
 import 'package:yak_error_handler/src/custom/error/wrong_handler.dart';
 import 'package:yak_error_handler/yak_error_handler.dart';
@@ -65,6 +63,21 @@ void main() {
           'ErrorHandler failed as'
           'an Error of type Error in place of AvowError',
           reason: '`toString` should be predicatble');
+    });
+
+    test(
+        'GIVEN `ErrorHandler<T>`'
+        'WHEN casted as `ErrorHandler<dynamc>` '
+        'THEN `type` still returs `T`', () {
+      final handler = ErrorHandler<AssertionError>((_) {});
+
+      final list = [handler];
+
+      expect(
+        list[0].type,
+        AssertionError,
+        reason: 'type should be persisted',
+      );
     });
   });
 }