+++ packages/yak_error_handler/test/extensions/handle_error/handle_error_from_error_test.dart
@@ -1,5 +1,5 @@
 import 'package:test/test.dart';
-import 'package:yak_runner/yak_runner.dart';
+import 'package:yak_error_handler/yak_error_handler.dart';
 
 void main() {
   group('`HandleErrorFromErrorExtension`', () {
@@ -7,7 +7,7 @@ void main() {
       'WHEN `T extends Error` & `T.handle()` THEN returns `HandleError<T>`',
       () => expect(
         AssertionError().handle((_) {}),
-        isA<HandleError<AssertionError>>(),
+        isA<ErrorHandler<AssertionError>>(),
         reason: '`HandleError.ignore()` should be `HandleError`',
       ),
     );
@@ -16,7 +16,7 @@ void main() {
       'WHEN `T extends Error` & `T.ignore` THEN returns `HandleError<T>`',
       () => expect(
         AssertionError().ignore,
-        isA<HandleError<AssertionError>>(),
+        isA<ErrorHandler<AssertionError>>(),
         reason: '`HandleError.ignore()` should be `HandleError`',
       ),
     );