+++ packages/yak_runner/lib/src/extensions/all.dart
@@ -1,3 +1,2 @@
 export 'custom/all.dart';
-export 'handle_exact_error/all.dart';
 export 'result/all.dart';