+++ packages/yak_runner/lib/src/extensions/handle_exact_error/from_error.dart
@@ -1,8 +1,0 @@
-import 'package:yak_error_handler/yak_error_handler.dart';
-
-/// an extension to easily create an `HandleError` from an `Error`
-extension HandleErrorFromErrorX<S extends Error> on S {
-  /// given a `void Function(T)` return a HandleErrorBase<T>
-  ErrorHandler<T, S> handle<T>(T? Function(Error) handleError) =>
-      ErrorHandler<T, S>(handleError);
-}