+++ packages/yak_runner/lib/src/runners/base.dart
@@ -1,17 +1,16 @@
 import 'package:yak_error_handler/yak_error_handler.dart';
-import 'package:yak_runner/yak_runner.dart';
 
 /// an interface for `yak_runner` class
 abstract class YakRunnerBase<T> {
   /// a function of typedef `HandleException`
-  final ExceptionHandler exceptionHandler;
+  final HandleExceptionDelegate? exceptionHandler;
 
   /// a list of specific to handle in a predetermined way
-  final Set<ErrorHandler<Failure<T>, dynamic>> errorHandlers;
+  final Set<ErrorHandler> errorHandlers;
 
   /// the constructor takes an optional parameter `errorHandler`
   const YakRunnerBase({
-    required this.exceptionHandler,
+    this.exceptionHandler,
     required this.errorHandlers,
   });
 }