+++ packages/yak_runner/test/mocks/delegates/handlers/all.dart
@@ -1,1 +1,2 @@
-export 'exception_handler.dart';
+export 'handle_error.dart';
+export 'handle_exception.dart';