+++ packages/yak_runner/test/mocks/delegates/handlers/handle_error.dart
@@ -1,0 +1,8 @@
+import 'package:stub/stub.dart';
+import 'package:yak_error_handler/yak_error_handler.dart';
+
+class HandleErrorDelegateStub extends Stub<void>
+    implements HandleErrorDelegate {
+  @override
+  void call(Error e) => stub();
+}