+++ packages/yak_runner/test/runners/async/unary_test.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 
 import 'package:test/test.dart';
+import 'package:yak_error_handler/yak_error_handler.dart';
 import 'package:yak_runner/yak_runner.dart';
 import '../../mocks/all.dart';
 //  ignore_for_file: avoid_catching_errors
@@ -9,24 +10,24 @@ void main() {
   const data = 1;
   const res = '$data';
   group('`YakRunnerArgAsync`', () {
-    final exceptionHandler = MockExceptionHandler();
-    final errorHandler = MockHandleError<AssertionError>();
-
+    final exceptionStub = HandleExceptionDelegateStub();
+    final errorStub = HandleErrorDelegateStub();
+    final errorHandler = ErrorHandler<AssertionError>(errorStub);
     final delegate = MockUnaryDelegate<Future<String>, FutureOr<int>>();
     final runner = YakRunnerArgAsync<String, int>(
       delegate,
-      handleException: exceptionHandler,
-      errorsWhitelist: {errorHandler},
+      exceptionHandler: exceptionStub,
+      errorHandlers: {errorHandler},
     );
 
     test('WHEN `void Function(T)` throws THEN `Result` is `Failure`', () async {
       delegate.reset;
-      exceptionHandler.reset;
-      errorHandler.reset;
+      exceptionStub.reset;
+      errorStub.reset;
 
       delegate.stub = () => throw Exception();
-      exceptionHandler.stub = () {};
-      errorHandler.stub = () {};
+      exceptionStub.stub = () {};
+      errorStub.stub = () {};
 
       final result = await runner(data);
 
@@ -51,12 +52,12 @@ void main() {
         reason: '`delegate` should be called once',
       );
       expect(
-        exceptionHandler.callCount,
+        exceptionStub.callCount,
         1,
-        reason: '`exceptionHandler` should be called once',
+        reason: '`exceptionStub` should be called once',
       );
       expect(
-        errorHandler.callCount,
+        errorStub.callCount,
         0,
         reason: '`errorHandler` should NOT be called',
       );
@@ -65,12 +66,12 @@ void main() {
     test('WHEN `void Function()` does not fail `Result` is `Success`',
         () async {
       delegate.reset;
-      exceptionHandler.reset;
-      errorHandler.reset;
+      exceptionStub.reset;
+      errorStub.reset;
 
       delegate.stub = () async => res;
-      exceptionHandler.stub = () {};
-      errorHandler.stub = () {};
+      exceptionStub.stub = () {};
+      errorStub.stub = () {};
 
       final result = await runner(data);
 
@@ -95,24 +96,24 @@ void main() {
         reason: '`delegate` should be called once',
       );
       expect(
-        exceptionHandler.callCount,
+        exceptionStub.callCount,
         0,
-        reason: '`exceptionHandler` should NOT be called ',
+        reason: '`exceptionStub` should NOT be called ',
       );
       expect(
-        errorHandler.callCount,
+        errorStub.callCount,
         0,
         reason: '`errorHandler` should NOT be called',
       );
     });
     test('WHEN `ERROR` is thwon THEN runner fails', () async {
       delegate.reset;
-      exceptionHandler.reset;
-      errorHandler.reset;
+      exceptionStub.reset;
+      errorStub.reset;
 
       delegate.stub = () => throw Error();
-      exceptionHandler.stub = () {};
-      errorHandler.stub = () {};
+      exceptionStub.stub = () {};
+      errorStub.stub = () {};
 
       Error? err;
 
@@ -133,12 +134,12 @@ void main() {
         reason: '`delegate` should be called once',
       );
       expect(
-        exceptionHandler.callCount,
+        exceptionStub.callCount,
         0,
-        reason: '`exceptionHandler` should NOT be called ',
+        reason: '`exceptionStub` should NOT be called ',
       );
       expect(
-        errorHandler.callCount,
+        errorStub.callCount,
         0,
         reason: '`errorHandler` should NOT be called',
       );
@@ -146,12 +147,12 @@ void main() {
 
     test('WHEN `AssertionError` is thwon THEN gets handled', () async {
       delegate.reset;
-      exceptionHandler.reset;
-      errorHandler.reset;
+      exceptionStub.reset;
+      errorStub.reset;
 
       delegate.stub = () => throw AssertionError();
-      exceptionHandler.stub = () {};
-      errorHandler.stub = () {};
+      exceptionStub.stub = () {};
+      errorStub.stub = () {};
 
       expect(
         await runner(data),
@@ -164,12 +165,12 @@ void main() {
         reason: '`delegate` should be called once',
       );
       expect(
-        exceptionHandler.callCount,
+        exceptionStub.callCount,
         0,
-        reason: '`exceptionHandler` should NOT be called ',
+        reason: '`exceptionStub` should NOT be called ',
       );
       expect(
-        errorHandler.callCount,
+        errorStub.callCount,
         1,
         reason: '`errorHandler` should be called once',
       );