+++ packages/yak_runner/test/runners/sync/nullary_test.dart
@@ -1,4 +1,5 @@
 import 'package:test/test.dart';
+import 'package:yak_error_handler/yak_error_handler.dart';
 import 'package:yak_runner/yak_runner.dart';
 import '../../mocks/all.dart';
 //  ignore_for_file: avoid_catching_errors
@@ -7,23 +8,24 @@ void main() {
   const data = 1;
 
   group('`YakRunner`', () {
-    final exceptionHandler = MockExceptionHandler();
+    final exceptionStub = HandleExceptionDelegateStub();
     final delegate = MockDelegate<int>();
-    final errorHandler = MockHandleError<AssertionError>();
+    final errorStub = HandleErrorDelegateStub();
+    final errorHandler = ErrorHandler<AssertionError>(errorStub);
     final runner = YakRunner(
       delegate,
-      handleException: exceptionHandler,
-      errorsWhitelist: {errorHandler},
+      exceptionHandler: exceptionStub,
+      errorHandlers: {errorHandler},
     );
 
     test('WHEN `void Function()` throws THEN `Result` is `Failure`', () {
       delegate.reset;
-      exceptionHandler.reset;
-      errorHandler.reset;
+      exceptionStub.reset;
+      errorStub.reset;
 
       delegate.stub = () => throw Exception();
-      exceptionHandler.stub = () {};
-      errorHandler.stub = () {};
+      exceptionStub.stub = () {};
+      errorStub.stub = () {};
 
       final result = runner();
 
@@ -48,12 +50,12 @@ void main() {
         reason: '`delegate` should be called once',
       );
       expect(
-        exceptionHandler.callCount,
+        exceptionStub.callCount,
         1,
-        reason: '`exceptionHandler` should be called once',
+        reason: '`exceptionStub` should be called once',
       );
       expect(
-        errorHandler.callCount,
+        errorStub.callCount,
         0,
         reason: '`errorHandler` should NOT be called',
       );
@@ -61,12 +63,12 @@ void main() {
 
     test('WHEN `void Function()` does not fail `Result` is `Success`', () {
       delegate.reset;
-      exceptionHandler.reset;
-      errorHandler.reset;
+      exceptionStub.reset;
+      errorStub.reset;
 
       delegate.stub = () => data;
-      exceptionHandler.stub = () {};
-      errorHandler.stub = () {};
+      exceptionStub.stub = () {};
+      errorStub.stub = () {};
 
       final result = runner();
 
@@ -91,12 +93,12 @@ void main() {
         reason: '`delegate` should be called once',
       );
       expect(
-        exceptionHandler.callCount,
+        exceptionStub.callCount,
         0,
-        reason: '`exceptionHandler` should NOT be called ',
+        reason: '`exceptionStub` should NOT be called ',
       );
       expect(
-        errorHandler.callCount,
+        errorStub.callCount,
         0,
         reason: '`errorHandler` should NOT be called',
       );
@@ -104,12 +106,12 @@ void main() {
 
     test('WHEN `ERROR` is thwon THEN runner fails', () {
       delegate.reset;
-      exceptionHandler.reset;
-      errorHandler.reset;
+      exceptionStub.reset;
+      errorStub.reset;
 
       delegate.stub = () => throw Error();
-      exceptionHandler.stub = () {};
-      errorHandler.stub = () {};
+      exceptionStub.stub = () {};
+      errorStub.stub = () {};
 
       Error? err;
 
@@ -130,12 +132,12 @@ void main() {
         reason: '`delegate` should be called once',
       );
       expect(
-        exceptionHandler.callCount,
+        exceptionStub.callCount,
         0,
-        reason: '`exceptionHandler` should NOT be called ',
+        reason: '`exceptionStub` should NOT be called ',
       );
       expect(
-        errorHandler.callCount,
+        errorStub.callCount,
         0,
         reason: '`errorHandler` should NOT be called',
       );
@@ -143,12 +145,12 @@ void main() {
     });
     test('WHEN `AssertionError` is thwon THEN gets handled', () {
       delegate.reset;
-      exceptionHandler.reset;
-      errorHandler.reset;
+      exceptionStub.reset;
+      errorStub.reset;
 
       delegate.stub = () => throw AssertionError();
-      exceptionHandler.stub = () {};
-      errorHandler.stub = () {};
+      exceptionStub.stub = () {};
+      errorStub.stub = () {};
 
       expect(
         runner(),
@@ -161,12 +163,12 @@ void main() {
         reason: '`delegate` should be called once',
       );
       expect(
-        exceptionHandler.callCount,
+        exceptionStub.callCount,
         0,
-        reason: '`exceptionHandler` should NOT be called ',
+        reason: '`exceptionStub` should NOT be called ',
       );
       expect(
-        errorHandler.callCount,
+        errorStub.callCount,
         1,
         reason: '`errorHandler` should be called once',
       );