+++ packages/yak_error_handler/CHANGELOG.md
@@ -1,6 +1,7 @@
-# 0.3.0+2
+# 0.3.0+3
 - minor update to `ErrorHandler<T,S>`
-- added internal `WrongErrorHandler`
+- added internal `WrongErrorHandler` and `from_error`
+- tests updated
 
 # 0.3.0+0
 - `NNBD`